requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
arrow==1.3.0
binaryornot==0.4.4
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cookiecutter==2.6.0
distro==1.9.0
flet==0.25.2
flet-cli==0.25.2
flet-desktop==0.25.2
groq==0.13.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.27
langchain-groq==0.2.2
langsmith==0.2.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
oauthlib==3.2.2
orjson==3.10.12
packaging==24.2
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
PyYAML==6.0.2
qrcode==7.4.2
repath==0.9.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
six==1.16.0
sniffio==1.3.1
style==1.1.0
tenacity==9.0.0
text-unidecode==1.3
toml==0.10.2
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
update==0.0.1
urllib3==2.2.3
watchdog==4.0.2
